requirements.txt
django
dj-database-url
psycopg2-binary
uwsgi
